separated Surface2D and Surface3D
