+++ course6/toyapp-adam/CMakeLists.txt
@@ -7,7 +7,7 @@ cmake_minimum_required(VERSION 3.0)
 
 set(SDL_ROOT "" CACHE STRING "SDL2 Library root path")
 
-if(NOT SDL_ROOT STREQUAL "")
+if(SDL_ROOT)
 	include_directories(${SDL_ROOT}/include)
 	link_directories(${SDL_ROOT}/lib)
 endif()
