+++ course6/toyapp-adam/displaylib/CMakeLists.txt
@@ -3,7 +3,8 @@ add_library(displaylib
 	color.h
 	cube.cc cube.h
 	sdlobject.cc sdlobject.h
-	surface.cc surface.h
+	surface_2d.cc surface_2d.h
+	surface_3d.cc surface_3d.h
 	window.cc window.h)
 
 target_link_libraries(displaylib SDL2)
